Vim: Simplified solution to 'n' jumping when not searching
I previously created a ridiculously elaborate solution to the problem
of "n" jumping when you were not searching. I discovered that the
:keeppatterns part of it was actually making itself not clear the
previous substitute search pattern.
Instead of keeping up this craziness, I figured out a far superior
solution. Now, "n" only jumps when search results are being highlit,
which lines up with when you are in the act of searching.